--- a/.travis.yml
+++ b/.travis.yml
@@ -39,8 +39,8 @@ env:
   - OPTIONS="CVR=ON MAKELIB=ON lammps" TEST="_lmp" CHOICES=gfort.lapack
   - LATTE_CMAKE=yes CMAKE_WITH_PROGRESS=yes
   - LATTE_CMAKE=yes LAMMPS_CMAKE=yes
-  - LATTE_PERFORMANCE=yes TEST="_lmp" CHOICES=gfort.lapack
-  - LATTE_PERFORMANCE=yes OPTIONS="CVR=ON PROGRESS=ON" TEST="" CHOICES=gfort.lapack
+  - LATTE_PERFORMANCE=yes TEST="" CHOICES=gfort.lapack
+#  - LATTE_PERFORMANCE=yes OPTIONS="CVR=ON PROGRESS=ON" TEST="" CHOICES=gfort.lapack
 
 script:
   -  cp -f -v ./makefiles/makefile.CHOICES.${CHOICES:-gfort.lapack} makefile.CHOICES