--- a/tests/run_test.sh
+++ b/tests/run_test.sh
@@ -6,12 +6,27 @@ MY_PATH=`pwd`                                   # Capturing the local path of th
 
 RUN="./LATTE_DOUBLE"  # LATTE program executable
 
-echo -e "\nTesting LATTE with new (latte.in) input files \n"
+echo ""
+echo "Reference elapsed CPU time"
+echo "=========================="
+echo ""
+
+( cd tests; make; cd .. ) &> /dev/null
+
+echo -en "   Testing for reference ... "
+timeRef=`( /usr/bin/time -f "%S" ./tests/timer_cpu_time > out ) 2>&1 > /dev/null`
+
+relativeTime=`echo "$timeRef/$timeRef" | bc -l | awk '{printf("%.3f",$1)}'`
+echo "(${timeRef}s,$relativeTime) PASSED"
+
+echo ""
+echo "Testing LATTE with new (latte.in) input files"
+echo "============================================="
+echo ""
 
 mv latte.in latte.in.tmp
 
 set -e                                          # This will exit the script if there is any error
-
 
 # Testing for single point calculations:
 
@@ -31,7 +46,8 @@ for name in single.point single.point.noelec single.point.rspace ; do
   ENERG=`grep -e "FREE ENERGY" out | awk 'NF>1{print $5}'`
   echo $ENERG > energy.out
 
-  echo -n "($time s) "
+  relativeTime=`echo "$time/$timeRef" | bc -l | awk '{printf("%.3f",$1)}'`
+  echo -n "(${time}s,$relativeTime) "
   python ./tests/test-energy.py --reference $REF --current energy.out --reltol 0.00001
 
   rm $REF out
@@ -57,7 +73,8 @@ for name in opt opt.cg opt_cons dorbitals; do
 
   time=`( /usr/bin/time -f "%S" $RUN > out ) 2>&1 > /dev/null`
 
-  echo -n "($time s) "
+  relativeTime=`echo "$time/$timeRef" | bc -l | awk '{printf("%.3f",$1)}'`
+  echo -n "(${time}s,$relativeTime) "
   python ./tests/test-optim.py --reference $REF --current monitorrelax.xyz --reltol 0.00001
 
   #rm $REF monitorrelax.xyz out
@@ -82,7 +99,8 @@ for name in tableread 0scf 2scf fullscf fullscf.etemp sp2 sp2.sparse fullscf.nvt
   
   grep "Data" out | sed 's/Data/ /g' | awk 'NF>1{print $2}' > energy.out
 
-  echo -n "($time s) "
+  relativeTime=`echo "$time/$timeRef" | bc -l | awk '{printf("%.3f",$1)}'`
+  echo -n "(${time}s,$relativeTime) "
   python ./tests/test-energy.py --reference $REF --current energy.out --reltol 0.00001
 
   rm $REF energy.out out
@@ -131,7 +149,8 @@ for name in fittingoutput.dat ; do
       }
     ' $REF`
   
-  echo -n "($time s) "
+  relativeTime=`echo "$time/$timeRef" | bc -l | awk '{printf("%.3f",$1)}'`
+  echo -n "(${time}s,$relativeTime) "
   if [ "$check" -eq 0 ]
   then
     echo "PASSED"
@@ -149,7 +168,10 @@ done
 
 rm latte.in
 
-echo -e "\nTesting LATTE with original input files \n"
+echo ""
+echo "Testing LATTE with original input files"
+echo "======================================="
+echo ""
 
 for name in 0scf fullscf sp2 ; do
 
@@ -170,7 +192,8 @@ for name in 0scf fullscf sp2 ; do
   grep "Data" out | sed 's/Data/ /g' | awk 'NF>1{print $2}' > energy.out
   grep Energy out | sed -e s/"PAR"/$STRR/g  >  input_tmp.in
   
-  echo -n "($time s) "
+  relativeTime=`echo "$time/$timeRef" | bc -l | awk '{printf("%.3f",$1)}'`
+  echo -n "(${time}s,$relativeTime) "
   python ./tests/test-energy.py --reference $REF --current energy.out --reltol 0.00001
 
   rm $REF energy.out input_tmp.in