Now the output of test looks as it is shown below. Each test includes the total user time and the relative execution time compared with a reference. If LATTE_PERFORMANCE=yes a test is considered that fails if the relative execution time is larger than 10% of the pre-established values described in the file test/run_test.sh

Reference elapsed CPU time
==========================

   Testing for reference ... (13.62s,1.000) PASSED

Testing LATTE with new (latte.in) input files
=============================================

   Testing for single.point ... (1.27s,0.093) PASSED
   Testing for single.point.noelec ... (0.65s,0.048) PASSED
   Testing for single.point.rspace ... (0.69s,0.051) PASSED
   Testing for opt ... (71.71s,5.265) PASSED
   Testing for opt.cg ... (75.30s,5.529) PASSED
   Testing for opt_cons ... (71.63s,5.259) PASSED
   Testing for dorbitals ... (48.83s,3.585) PASSED
   Testing for tableread ... (12.43s,0.913) PASSED
   Testing for 0scf ... (8.78s,0.645) PASSED
   Testing for 2scf ... (7.68s,0.564) PASSED
   Testing for fullscf ... (11.66s,0.856) PASSED
   Testing for fullscf.etemp ... (11.24s,0.825) PASSED
   Testing for sp2 ... (6.95s,0.510) PASSED
   Testing for sp2.sparse ... (37.78s,2.774) PASSED
   Testing for fullscf.nvt ... (12.39s,0.910) PASSED
   Testing for fullscf.npt ... (12.44s,0.913) PASSED
   Testing for fullscf.vdw ... (11.54s,0.847) PASSED
   Testing for fullscf.spin ... (16.14s,1.185) PASSED
   Testing for fullscf.kon ... (30.04s,2.206) PASSED
   Testing for fullscf.rspace ... (3.72s,0.273) PASSED
   Testing for fittingoutput.dat ... (1.00s,0.073) PASSED

Testing LATTE with original input files
=======================================

   Testing for 0scf ... (8.14s,0.598) PASSED
   Testing for fullscf ... (13.64s,1.001) PASSED
   Testing for sp2 ... (7.32s,0.537) PASSED

End of run and test