@cnegre 2018-07-06
That's what he is doing. The first test is a dummy test which is used as
the time unit.
We could check is the workload of a node could change these relative times.

On Fri, Jul 6, 2018 at 4:35 PM, Christoph Junghans <notifications@github.com
> wrote:

> *@junghans* commented on this pull request.
> ------------------------------
>
> In tests/run_test.sh
> <https://github.com/lanl/LATTE/pull/75#discussion_r200785622>:
>
> > @@ -4,6 +4,32 @@
>
>  MY_PATH=`pwd`                                   # Capturing the local path of the folder where we are running.
>
> +declare -A performanceExpectedTimes
> +performanceExpectedTimes["single.point"]=0.014
>
> Or make it relative to the 1st test!
>
> —
> You are receiving this because you modified the open/close state.
> Reply to this email directly, view it on GitHub
> <https://github.com/lanl/LATTE/pull/75#discussion_r200785622>, or mute
> the thread
> <https://github.com/notifications/unsubscribe-auth/APQltcsurFTQkeruN9fdpqfHZMVyzu00ks5uD-ZAgaJpZM4VB_wj>
> .
>



-- 
Dr. Christian F. A . Negre

LOS ALAMOS NATIONAL LABORATORY
T-1 Division, Mail Stop B221
Los Alamos, NM 87545