@nfaguirrec 2018-07-10
Regarding the false positives in the performance test, I think we do not have that problem. I am using the "user time (U)". According to the manual of the command "time", we have three possible options:

E      Elapsed real (wall clock) time used by the process, in [hours:]minutes:seconds.
S      Total number of CPU-seconds used by the system on behalf of the process (in kernel mode), in seconds.
U      Total number of CPU-seconds that the process used directly (in user mode), in seconds.

So, if I understood well, in principle, the option U should be independent of the load on the host but maybe I am wrong. By using the option S I got weird results and actually, I don't understand pretty well what kernel mode means.